0.02
First playable version! ball now bounces naturally, random angle
assigned at beginning, etc.

+++ Tau/player.h
@@ -9,7 +9,7 @@ class Player : public QObject, public QGraphicsItem
     Q_INTERFACES(QGraphicsItem)
 public:
     Player(double pos, double size,int playerNum);
-    double pos; // angular position
+    int pos; // angular position
     double size; // angular distance from center to edge
     int startAngle();
     int spanAngle();
@@ -29,7 +29,6 @@ protected:
     QBrush brush;
     QRectF* rekt;
     void paint ( QPainter * painter, const QStyleOptionGraphicsItem * option, QWidget * widget);
-
 };
 
 #endif // PLAYER_H
